update to new package

+++ install.ps1
@@ -1,5 +1,6 @@
 
 # TODO: Should allow user to opt-out of installing python3 and use a pre-installed version
+# TODO: Currently requires outbound internet to fetch python 3.7.7 and update pip
 
 # Expects to be run from inside AUTOSCALER_HOME
 # Expects to find nuget on the PATH or in JETPACK_BIN
@@ -8,18 +9,22 @@
 [string]$BOOTSTRAP = "C:\cycle\jetpack\system\bootstrap"
 [string]$AUTOSCALER_VENV="c:\cycle\hpcpack-autoscaler\.venvs\cyclecloud-hpcpack"
 
-mkdir "c:\cycle"
+mkdir "$AUTOSCALER_HOME"
 
 $env:Path += ";" + $JETPACK_BIN
-nuget.exe install python -Version 3.7.7 -OutputDirectory C:\cycle\
-& C:\cycle\python.3.7.7\tools\python.exe -m venv $AUTOSCALER_VENV
+if (-not (Test-Path "C:\cycle\python.3.7.7")) {
+    nuget.exe install python -Version 3.7.7 -OutputDirectory C:\cycle\
+}
+if (-not (Test-Path "$AUTOSCALER_VENV")) {
+    & C:\cycle\python.3.7.7\tools\python.exe -m venv $AUTOSCALER_VENV
+}
 & $AUTOSCALER_VENV\Scripts\Activate.ps1
 
 
 & pip install -U pip
 # & pip install urllib3 requests typeguard jsonpickle pytz
-& pip install packages\*
-& pip install -e .
+& pip install -U packages\*
+# & pip install -e .
 
 
 @"
